Use decodeURI() to replace %20 with spaces
When the folder to download contains special characters (e.g. space :-)) in its name, the ZIP file name and the name of the root folder in the file are URL encoded. This makes for ugly file names like "Joe%20Cocker" instead of "Joe Cocker". This patch wraps the corresponding variables with decodeURI() to get pretty file names.